Lint shell scripts and markdown files

+++ test-scripts/helper/cleanup.sh
@@ -8,12 +8,12 @@ source ./helper/var.env
 
 # cleanup step
 echo "Ensuring no container mounted in mount directory"
-sudo fusermount -u $MOUNT_DIR
-sudo fusermount3 -u $MOUNT_DIR
+sudo fusermount -u "$MOUNT_DIR"
+sudo fusermount3 -u "$MOUNT_DIR"
 
 echo "Stopping previous run of cloudfuse"
-sudo kill -9 `pidof cloudfuse` || true
+sudo kill -9 "$(pidof cloudfuse)" || true
 
 echo "Deleting files in mount and temp directories"
-rm -rf $MOUNT_DIR/*
-rm -rf $MOUNT_TMP/*
+rm -rf "${MOUNT_DIR:?}/*"
+rm -rf "${MOUNT_TMP:?}/*"
